tokens

+++ Leftovers/Leftovers/Controllers/MealsController.cs
@@ -3,7 +3,9 @@ using Leftovers.Data.Repositories;
 using Leftovers.Data.Entities;
 using Leftovers.Data.Dtos.Meals;
 using AutoMapper;
-
+using Microsoft.AspNetCore.Authorization;
+using Leftovers.Auth.Model;
+using System.Security.Claims;
 namespace Leftovers.Controllers
 {
     [ApiController]
@@ -14,12 +16,14 @@ namespace Leftovers.Controllers
         private readonly IMapper _mapper;
         private readonly IRestaurantsRepository _restaurantsRepository;
         private readonly IChainsRepository _chainsRepository;
-        public MealsController(IMealsRepository mealsRepository, IRestaurantsRepository restaurantsRepository, IChainsRepository chainsRepository, IMapper mapper)
+        private readonly IAuthorizationService _authorizationService;
+        public MealsController(IMealsRepository mealsRepository, IRestaurantsRepository restaurantsRepository, IChainsRepository chainsRepository, IMapper mapper, IAuthorizationService authorizationService)
         {
             _mealsRepository = mealsRepository;
             _mapper = mapper;
             _chainsRepository = chainsRepository;
             _restaurantsRepository = restaurantsRepository;
+            _authorizationService = authorizationService;
         }
         [HttpGet]
         public async Task<IEnumerable<MealDto>> GetAllAsync(int restaurantId)
@@ -36,6 +40,7 @@ namespace Leftovers.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = LeftoversUserRoles.RestaurantUser)]
         public async Task<ActionResult<MealDto>> PostAsync(int chainId, int restaurantId, CreateMealDto mealDto)
         {
             var chain = await _chainsRepository.GetAsync(chainId);
@@ -47,10 +52,14 @@ namespace Leftovers.Controllers
             var meal = _mapper.Map<Meal>(mealDto);
             meal.RestaurantId = restaurantId;
             await _mealsRepository.InsertAsync(meal);
+
+            meal.UserId = User.FindFirstValue(CustomClaims.UserId);
+
             return Created($"api/chain/{chainId}/restaurant/{restaurantId}/meal/{meal.Id}", _mapper.Map<MealDto>(meal));
         }
 
         [HttpPut("{mealId}")]
+        [Authorize(Roles = LeftoversUserRoles.RestaurantUser)]
         public async Task<ActionResult<MealDto>> PutAsync(int chainId, int restaurantId, int mealId, UpdateMealDto mealDto)
         {
             var chain = await _mealsRepository.GetAsync(chainId);
@@ -62,6 +71,13 @@ namespace Leftovers.Controllers
             var oldMeal = await _mealsRepository.GetAsync(restaurantId, mealId);
             if (oldMeal == null) return NotFound($"Chain with id '{chainId}' not found");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, oldMeal, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();//403
+            }
+
+
             _mapper.Map(mealDto, oldMeal);
             await _mealsRepository.UpdateAsync(oldMeal);
             return Ok(_mapper.Map<MealDto>(oldMeal));
@@ -69,10 +85,18 @@ namespace Leftovers.Controllers
         
 
         [HttpDelete("{mealId}")]
+        [Authorize(Roles = LeftoversUserRoles.RestaurantUser)]
         public async Task<ActionResult<Meal>> DeleteAsync(int restaurantId, int mealId)
         {
             var meal = await _mealsRepository.GetAsync(restaurantId, mealId);
             if (meal == null) return NotFound($"Meal with id '{mealId}' not found");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, meal, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();//403
+            }
+
             await _mealsRepository.DeleteAsync(meal);
             //204
             return NoContent();
